major refactor
not done, needs commenting and adjustments and namespace changes
probably

* refactored the raycasting system. instead of tilemarching, we use
robust engine's rays with predicate. rays also bounce and come off of
the player instead of audio sources (achieves the same effect for much
cheaper)

* refactored the way echo is calculated. rays will keep track of
entities that have the new client-side component
AudioAbsorptionComponent (probably gonna rename it to like...
AudioAcousticsComponent or something). averages out all the absorption
values it finds and does coefficient stuff and i still need to tweak
this a bit

* ignore the blood stains